fix: add metadataUtils + add guard before parsing metadata
