+++ client/src/app/components/stardust/address/section/nft/Nft.tsx
@@ -5,7 +5,7 @@ import nftSchemeIRC27 from "../../../../../../assets/schemas/nft-schema-IRC27.js
 import { useNftMetadataUri } from "../../../../../../helpers/hooks/useNftMetadataUri";
 import { useTokenRegistryNftCheck } from "../../../../../../helpers/hooks/useTokenRegistryNftCheck";
 import { Bech32AddressHelper } from "../../../../../../helpers/stardust/bech32AddressHelper";
-import { tryParseMetadata } from "../../../../../../helpers/stardust/valueFormatHelper";
+import { tryParseMetadata } from "../../../../../../helpers/stardust/metadataUtils";
 import { INftImmutableMetadata } from "../../../../../../models/api/stardust/nft/INftImmutableMetadata";
 import NetworkContext from "../../../../../context/NetworkContext";
 import TruncatedId from "../../../TruncatedId";
