+++ client/src/app/components/stardust/address/section/nft/NftMetadataSection.tsx
@@ -3,7 +3,7 @@ import { Link } from "react-router-dom";
 import nftSchemeIRC27 from "../../../../../../assets/schemas/nft-schema-IRC27.json";
 import { useNftMetadataUri } from "../../../../../../helpers/hooks/useNftMetadataUri";
 import { useTokenRegistryNftCheck } from "../../../../../../helpers/hooks/useTokenRegistryNftCheck";
-import { tryParseMetadata } from "../../../../../../helpers/stardust/valueFormatHelper";
+import { tryParseMetadata } from "../../../../../../helpers/stardust/metadataUtils";
 import { INftBase } from "../../../../../../models/api/stardust/nft/INftBase";
 import { INftImmutableMetadata } from "../../../../../../models/api/stardust/nft/INftImmutableMetadata";
 import DataToggle from "../../../../DataToggle";
