+++ client/src/helpers/stardust/metadataUtils.ts
@@ -1,0 +1,22 @@
+import { HexEncodedString } from "@iota/iota.js-stardust";
+import { Converter } from "@iota/util.js-stardust";
+import * as jsonschema from "jsonschema";
+/**
+ * Tries to parse hex data into metadata.
+ * @param metadataHex The encoded data.
+ * @param schema The json schema to validate aginst.
+ * @returns The parsed metadata or undefined.
+ */
+export function tryParseMetadata<S>(metadataHex: HexEncodedString, schema: jsonschema.Schema): S | null {
+    const validator = new jsonschema.Validator();
+    try {
+        const json: unknown = JSON.parse(Converter.hexToUtf8(metadataHex));
+        const result = validator.validate(json, schema);
+
+        if (result.valid) {
+            return json as S;
+        }
+    } catch { }
+
+    return null;
+}
