+++ client/src/helpers/stardust/valueFormatHelper.tsx
@@ -1,6 +1,4 @@
-import { HexEncodedString, INodeInfoBaseToken, UnitsHelper } from "@iota/iota.js-stardust";
-import { Converter } from "@iota/util.js-stardust";
-import * as jsonschema from "jsonschema";
+import { INodeInfoBaseToken, UnitsHelper } from "@iota/iota.js-stardust";
 import React from "react";
 import Tooltip from "../../app/components/Tooltip";
 /**
@@ -61,24 +59,3 @@ export function formatSpecialBlockId(id: string): React.ReactNode {
     }
     return id;
 }
-
-/**
- * Tries to parse hex data into metadata.
- * @param metadataHex The encoded data.
- * @param schema The json schema to validate aginst.
- * @returns The parsed metadata or undefined.
- */
-export function tryParseMetadata<S>(metadataHex: HexEncodedString, schema: jsonschema.Schema): S | null {
-    const validator = new jsonschema.Validator();
-    try {
-        const json: unknown = JSON.parse(Converter.hexToUtf8(metadataHex));
-        const result = validator.validate(json, schema);
-
-        if (result.valid) {
-            return json as S;
-        }
-    } catch { }
-
-    return null;
-}
-
